Assignment/auth

+++ crud-app/app/Http/Controllers/ProductController.php
@@ -24,8 +24,13 @@ class ProductController extends Controller
      * Show the form for creating a new resource.
      */
 
-    public function create()
+    public function create(Request $request)
     {
+
+        if ($request->user()->cannot('create', Product::class)) {
+            return redirect()->route('products.index')->with('error', 'You do not have permission');
+        };
+
         $product = new Product;
         return view('products.create', ['product' => $product]);
     }
@@ -52,8 +57,12 @@ class ProductController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit($id)
+    public function edit($id, Request $request)
     {
+        if ($request->user()->cannot('update', Product::class)) {
+            return redirect()->route('products.index')->with('error', 'You do not have permission');
+        }
+
         $product =  Product::findOrFail($id);
         return view('products.edit', ['product' => $product]);
     }
@@ -63,15 +72,23 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if ($request->user()->cannot('update', Product::class)) {
+            return redirect()->route('products.index')->with('error', 'You do not have permission');
+        };
+
         Product::findOrFail($id)->update($this->validatedData($request));
         return redirect()->route('products.index')->with('success', 'Product was updated');
-    }
+                             }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($id)
+    public function destroy($id, Request $request)
     {
+        if ($request->user()->cannot('delete', Product::class)) {
+            return redirect()->route('products.index')->with('error', 'You do not have permission');
+        };
+
         $product = Product::findOrFail($id);
         $product->delete();
         return redirect()->route('products.index')->with('success', 'Product was deleted');
